Added ability to write XiaData to char array
Fixed several errors in writing XiaData to disk.

Changed XiaData::energy to unsigned int and time
and eventTime to unsigned long long to make it
easier to compare times, write to file, and to
prevent rounding errors.

Added the ability to write XiaData events to a
character array as well as an ofstream.