Modified the way CFD is performed by ChannelEvent
Added a new CFD algorithm to ChannelEvent class in
order to replace the XIA-style algorithm which was
used previously. This is done to increase timing
resolution while significantly reducing the amount
of tuning which is needed to measure high resolution
time differences.

ChannelEvent class no longer fills the cfd value
array when ::AnalyzeCFD is called. Scope program
now only draws the phase line when cfd is active.