Changed the way that scope saves traces to file
Scope now has the ability to save average traces and
profiles to file instead of just saving individual
traces.

Changed the way that the "save" command works in the
scope program. This command will now save the trace
or average trace histogram (and TProfile) that is
currently displayed to the screen. This is useful
when the user stops the acquisition on a trace
which they want to save.

Previously, the scope program would save whatever
trace which would be displayed after the user
typed the "save" command.