Added ability to specify scan stop point
Added the ability to specify the stop point in the
input file. The scan will continue until the position
in the input file is greater than or equal to the user
specified stop word (or percentage).

Added the ability to specify the starting input file
offset as a percentage (.xx) instead of specifying
it in 4 byte words.